Add :description to required fields on organisation changesets.
As mentioned in 5b951c1a510c86e4877b18c0766fc65c290c9e53 did 5e6b235c3078f9dbacfa5295e27b5b3fc770a1d0 introduce a not null contraint on the organization description.

This adds the description to the required fields on the organization changesets so it is possible to catch the constraint early.