Fixes log printing in non-hosted environments
This changes the way in which the Logger is configured to integrate with
the Timber package. It inverts the prior determination of whether to use
the "machine-friendly" version of the Timber output.

"Environment" here refers to the Mix build environment which CodeCorps
also uses for runtime environment.

In the existing implmentation, the "machine-friendly" output is used for
every environment except `:test`. It should really only be used for
hosted environments. At the moment, those environments are `:prod` and
`:staging`.

The logging configuration will now use the "developer-friendly" version
of Timber output by default. If the environment is `:prod` or `:staging`
it will switch to the "machine-friendly" version of Timber output.

I also removed `:timber` from the `:extra_applications` field in the Mix
configuration. Elixir will automatically include `:timber` in the
compiled applications list based on the dependencies.